requirements.txt
asyncpraw
langgraph
langchain-core
transformers
pandas
numpy
praw
plotly
python-dotenv
streamlit

